v17 work in progress

+++ SFUI-Whorehouse/Download.cpp
@@ -10,19 +10,19 @@ namespace fs = std::experimental::filesystem;
 
 Download::Download()
 {
-	std::cout << "downloader created" << std::endl;
+	std::cout << "[DL] downloader created" << std::endl;
 }
 
 Download::~Download()
 {
-	std::cout << "downloader killed" << std::endl;
+	std::cout << "[DL] downloader killed" << std::endl;
 }
 
 void Download::setOutputDir(std::string dir)
 {
 	saveDir = dir;
 
-	std::cout << "output set to: " << saveDir << std::endl;
+	std::cout << "[DL] output set to: " << saveDir << std::endl;
 
 	if (!fs::exists(saveDir))
 		createDirectory(saveDir);
@@ -37,7 +37,7 @@ void Download::setOutputFilename(std::string file)
 {
 	saveFile = file;
 
-	std::cout << "output file set to: " << saveFile << std::endl;
+	std::cout << "[DL] output file set to: " << saveFile << std::endl;
 }
 
 std::string Download::getOutputFilename()
@@ -47,17 +47,13 @@ std::string Download::getOutputFilename()
 
 void Download::setInput(std::string in)
 {
-//	std::cout << "raw input path: " << in << std::endl;
-
-	input = in;
-
-	std::string temppath = input;
+	std::string temppath = in;
 
 	for (size_t i = 0; i < temppath.length(); i++)
 		if (temppath[i] == '//')
 			temppath[i] = '/'; // change all the slashes to forwards so that FTP accepts them
 
-	std::cout << "steralised input path: " << temppath << std::endl;
+	std::cout << "[DL] sterilized input path: " << temppath << std::endl;
 
 	std::string filename = temppath;
 	std::string::size_type pos = filename.find_last_of("/\\");
@@ -66,9 +62,6 @@ void Download::setInput(std::string in)
 	remoteFilename = filename;
 
 	remoteDirectory = temppath.erase(temppath.length() - remoteFilename.length(), temppath.back());
-
-//	std::cout << "remoteDirectory: " << remoteDirectory << std::endl;
-//	std::cout << "remoteFilename : " << remoteFilename << std::endl;
 }
 
 std::string Download::getInput()
@@ -96,32 +89,32 @@ std::string Download::getInputFilename()
 	return remoteFilename;
 }
 
-// TODO: make this function useable, by adding some sort of decryption thingy
 uintmax_t Download::getFileSize()
 {
+// TODO: make this function useable, by adding some sort of decryption thingy
+	sf::Ftp ftp;
+	
 	if (!downloaded && !remoteFilename.empty())
 	{
-		std::cout << "retrieving file size" << std::endl;
+		std::cout << "[DL] retrieving file size" << std::endl;
 
 		// Connect to the server
 		sf::Ftp::Response response = ftp.connect("files.000webhost.com");
 		if (response.isOk())
-			std::cout << "Connected" << std::endl;
+			std::cout << "[DL] Connected" << std::endl;
 
 		// Log in
 		response = ftp.login("kunlauncher", "9fH^!U2=Ys=+XJYq");
 		if (response.isOk())
-			std::cout << "Logged in" << std::endl;
+			std::cout << "[DL] Logged in" << std::endl;
 
-		std::cout << "remote directory + remote filename: " << remoteDirectory + remoteFilename << std::endl;
+		std::cout << "[DL] remote directory + remote filename: " << remoteDirectory + remoteFilename << std::endl;
 
 		response = ftp.sendCommand("SIZE", remoteDirectory + remoteFilename);
 		if (response.isOk())
-			std::cout << "File size: " << response.getMessage() << std::endl;
+			std::cout << "[DL] File size: " << response.getMessage() << std::endl;
 		else
-		{
-			std::cout << response.getMessage() << std::endl;
-		}
+			std::cerr << response.getMessage() << std::endl;
 
 		ftp.disconnect();
 
@@ -145,31 +138,37 @@ int Download::download()
 
 	sf::Ftp ftp;
 
+	downloaded = false;
+	sizeVerified = false;
+
 	sf::Ftp::Response response = ftp.connect("files.000webhost.com", 21, sf::milliseconds(10000));
 	if (!response.isOk())
 	{
-		std::cerr << "failed to connect to ftp (" << response.getMessage() << " (" << response.getStatus() << "))" << std::endl;
-		return Status::ConnectionFailed;
+		std::cerr << "[DL] failed to connect to ftp (" << response.getMessage() << " (" << response.getStatus() << "))" << std::endl;
+		return Status::Fail | Status::ConnectionFailed;
 	}
 
 	response = ftp.login("kunlauncher", "9fH^!U2=Ys=+XJYq");
 	if (!response.isOk())
 	{
-		std::cout << "failed to connect to ftp" << std::endl;
-
-		return Status::Failure;
+		std::cout << "[DL] failed to login to ftp" << std::endl;
+		return Status::Fail | Status::ConnectionRejected;
 	}
 
 	response = ftp.changeDirectory("public_html");
 	if (!response.isOk())
 	{
-		std::cerr << "failed to set ftp directory" << std::endl;
-		return Status::Failure;
+		std::cerr << "[DL] failed to set ftp directory" << std::endl;
+		return Status::Fail;
 	}
 
+	// TODO: sometimes this, we fail to download and get 500 RETR.
+	// since the download happens after this, there will be no resource
+	// manifest until a new one is downloaded (usually after a restart)
+	// SOLUTION: instead, we should rename the file and then delete it later
 	if (fs::exists(GBL::DIR::cache + remoteDirectory + remoteFilename))
 	{
-		std::cout << "file already exists in cache, removing." << std::endl;
+		std::cout << "[DL] file already exists in cache, removing." << std::endl;
 
 		try
 		{
@@ -177,7 +176,7 @@ int Download::download()
 		}
 		catch (const std::exception& e)
 		{
-			std::cerr << "failed to remove already existing file" << std::endl;
+			std::cerr << "[DL] failed to remove already existing file" << std::endl;
 			std::cerr << e.what() << std::endl;
 		}
 	}
@@ -185,61 +184,97 @@ int Download::download()
 	if (!fs::exists(GBL::DIR::cache + remoteDirectory))
 		createDirectory(GBL::DIR::cache + remoteDirectory);
 
-	response = ftp.download(remoteDirectory + remoteFilename, GBL::DIR::cache + remoteDirectory);
+	// we use this to verify that the download was at least full
+	response = ftp.sendCommand("SIZE", remoteDirectory + remoteFilename);
 	if (response.isOk())
 	{
-		std::cout << response.getStatus() << ": downloaded the thing to " << saveDir << std::endl;
+		std::cout << "[DL] remote file size: " << response.getMessage() << std::endl;
+		fileSize = std::stoi(response.getMessage());
+	}
+	else
+	{
+		std::cerr << "[DL] (" << response.getStatus() << ") " << response.getMessage() << std::endl;
+	}
+
+	std::cout << "[DL] BEGIN DOWNLOAD" << std::endl;
+
+	response = ftp.download(remoteDirectory + remoteFilename, GBL::DIR::cache + remoteDirectory, sf::Ftp::TransferMode::Binary);
+	if (response.isOk())
+	{
+		std::cout << "[DL] " << response.getStatus() << ": download success, saved to " << saveDir << std::endl;
 
 		if (saveFile.empty())
 			saveFile = remoteFilename;
 
-		std::ifstream fileContent(".//bin//cache//" + remoteDirectory + remoteFilename, std::ios::binary);
-		if (fileContent)
+		// TODO: possibly don't overwrite file
+		std::ifstream fileContent(GBL::DIR::cache + remoteDirectory + remoteFilename, std::ios::binary, std::ios::trunc);
+		if (fileContent.is_open())
 		{
 			fileBuffer = (std::string((std::istreambuf_iterator<char>(fileContent)), std::istreambuf_iterator<char>()));
 
-			std::cout << "downloaded file, wrote to buffer: " << getAppropriateFileSize(fileBuffer.size(), 3) << std::endl;
+			std::cout << "[DL] wrote file to to buffer: " << getAppropriateFileSize(fileBuffer.size(), 3) << std::endl;
 			fileSize = fileBuffer.size();
+
+			fileContent.close();
+
+			// verify file size
+			if (fileSize != fs::file_size(GBL::DIR::cache + remoteDirectory + remoteFilename))
+			{
+				std::cerr << "[DL] downloaded file size differs from file size reported by server!" << std::endl;
+				sizeVerified = true;
+			}
 		}
 		else
 		{
-			std::cerr << "failed to open saved file" << std::endl;
+			std::cerr << "[DL] failed to open file for reading into buffer" << std::endl;
+			return Status::Fail;
 		}
-		fileContent.close();
 	}
 	else
 	{
-		std::cerr << response.getMessage() << ": something went wrong! (" << response.getStatus() << ")" << std::endl;
-		return Status::Failure;
+		std::cerr << "[DL] " << "(" << response.getStatus() << ")" << response.getMessage() << std::endl;
+		return Status::Fail;
 	}
 
 	ftp.disconnect();
-
 	downloaded = true;
-	return Status::Ok;
+
+	return Status::Success;
 }
 
 int Download::save()
 {
 	try
 	{
+		if (!fs::exists(GBL::DIR::cache + remoteDirectory + remoteFilename))
+		{
+			std::cerr << "[DL] downloaded file does not exist in cache" << std::endl;
+			return Status::Fail;
+		}
+
+		if (!downloaded)
+		{
+			std::cerr << "[DL] download status was not set as true!" << std::endl;
+			return Status::Fail;
+		}
+
 		if (fs::exists(saveDir + saveFile))
 		{
-			std::cout << "output file already exists, overwriting" << std::endl;
+			std::cout << "[DL] output file already exists, overwriting" << std::endl;
 			fs::remove(saveDir + saveFile);
 		}
 
 		fs::copy_file(GBL::DIR::cache + remoteDirectory + remoteFilename, saveDir + saveFile);
-		std::cout << "saved file" << std::endl;
+		std::cout << "[DL] saved file" << std::endl;
 
-		return Status::Ok;
+		return Status::Success;
 	}
 	catch (const std::exception &e)
 	{
-		std::cerr << "failed to save file:" << std::endl;
+		std::cerr << "[DL] failed to save file:" << std::endl;
 		std::cerr << e.what() << std::endl;
 
-		return Status::SaveFailed;
+		return Status::Fail;
 	}
 }
 
@@ -248,6 +283,7 @@ std::string Download::getAppropriateFileSize(const long long int bytes, const in
 	int bytesPerUnit = 1024;
 	std::string sizes[] = { "Bytes", "KB", "MB", "GB", "TB", "PB", "EB", "ZB", "YB", "REALLY FUCKIN BIG" };
 	int i = std::floor(std::log(bytes) / std::log(bytesPerUnit));
+	// TODO: look into changing int to double
 
 	std::string number = std::to_string(bytes / std::pow(bytesPerUnit, i));
 	number.erase(number.find('.') + decimals + 1, number.length());
@@ -259,11 +295,11 @@ void Download::clearCache()
 	try
 	{
 		fs::remove_all(GBL::DIR::cache);
-		std::cout << "download cache cleared" << std::endl;
+		std::cout << "[DL] upload cache cleared" << std::endl;
 	}
 	catch (const std::exception& e)
 	{
-		std::cerr << "failed to clear download cache:" << std::endl;
+		std::cerr << "[DL] failed to clear upload cache:" << std::endl;
 		std::cerr << e.what() << std::endl;
 	}
 }
@@ -272,7 +308,7 @@ void Download::clearCache()
 
 void Download::createDirectory(std::string dir) // recursively
 {
-	std::cout << "creating " << dir << std::endl;
+	std::cout << "[DL] creating " << dir << std::endl;
 
 	std::vector<fs::path> subdirectories;
 //	dir.erase(0, 2); // .//
